Move test-only method to call site
