+++ client/trino-jdbc/src/test/java/io/trino/jdbc/TestAsyncResultIterator.java
@@ -29,6 +29,7 @@ import org.junit.jupiter.api.Timeout;
 
 import java.net.URI;
 import java.time.ZoneId;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -41,7 +42,6 @@ import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.Supplier;
 
-import static io.trino.client.ResultRows.wrapList;
 import static java.util.Objects.requireNonNull;
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -64,7 +64,7 @@ class TestAsyncResultIterator
                     catch (InterruptedException e) {
                         interruptedButSwallowedLatch.countDown();
                     }
-                    return wrapList(ImmutableList.of(ImmutableList.of(new Object())));
+                    return fromList(ImmutableList.of(ImmutableList.of(new Object())));
                 }), ignored -> {},
                 new WarningsManager(),
                 Optional.of(new ArrayBlockingQueue<>(100)));
@@ -93,7 +93,7 @@ class TestAsyncResultIterator
         AsyncResultIterator iterator = new AsyncResultIterator(
                 new MockStatementClient(() -> {
                     thread.compareAndSet(null, Thread.currentThread());
-                    return wrapList(ImmutableList.of(ImmutableList.of(new Object())));
+                    return fromList(ImmutableList.of(ImmutableList.of(new Object())));
                 }), ignored -> {},
                 new WarningsManager(),
                 Optional.of(queue));
@@ -368,6 +368,26 @@ class TestAsyncResultIterator
             public Long getUpdateCount()
             {
                 throw new UnsupportedOperationException();
+            }
+        };
+    }
+
+    static ResultRows fromList(List<List<Object>> values)
+    {
+        return new ResultRows() {
+            @Override
+            public void close() {}
+
+            @Override
+            public Iterator<List<Object>> iterator()
+            {
+                return values.iterator();
+            }
+
+            @Override
+            public String toString()
+            {
+                return "ResultRows{values=" + values + "}";
             }
         };
     }
